requirements.txt
av==10.0.0
Pillow==10.0.1
PyYAML==5.3.1
